Добавлена реализация двенадцатого задания шестой главы
