+++ source/js/3d-animations/ProjectGui/TransformationGuiHelper.js
