+++ source/js/3d-animations/controllers/AnimationManager.js
