+++ source/js/3d-animations/creators/ExtrudeSvgCreator.js
