+++ source/js/3d-animations/helpers/ColorGuiHelper.js
