+++ source/js/3d-animations/initAnimationScreen.js
