+++ source/js/3d-animations/loaders/SvgPathLoader.js
