+++ source/js/3d-animations/mesh-complex-objects/Carpet.js
