+++ source/js/Animation/animation.js
@@ -1,0 +1,113 @@
+import {easeLinear} from "../helpers/easing";
+
+export default class Animation {
+  constructor(options) {
+    this.options = options;
+
+    if (!this.options.easing) {
+      this.options.easing = easeLinear;
+    }
+
+    if (!this.options.duration) {
+      this.options.duration = 1000;
+    }
+
+    if (!this.options.delay) {
+      this.options.delay = 0;
+    }
+
+    if (!this.options.fps) {
+      this.options.fps = 60;
+    }
+
+    this.timeoutId = null;
+    this.requestId = null;
+  }
+
+  start(options) {
+    this.stop();
+
+    this.timeoutId = setTimeout(() => {
+      this.startTime = performance.now();
+      this.interval = 1000 / this.options.fps;
+      this.lastFrameTime = this.startTime;
+      this.isFinished = false;
+
+      let animateFrame;
+
+      if (this.options.duration === "infinite") {
+        animateFrame = (currentTime) => {
+          const delta = currentTime - this.lastFrameTime;
+
+          if (delta > this.interval) {
+            this.options.func(1, {
+              startTime: this.startTime,
+              currentTime,
+              isFinished: false,
+              options,
+            });
+
+            this.lastFrameTime = currentTime;
+          }
+
+          this.requestId = requestAnimationFrame(animateFrame);
+        };
+      } else {
+        animateFrame = (currentTime) => {
+          const delta = currentTime - this.lastFrameTime;
+
+          if (delta > this.interval) {
+            let timeFraction =
+              (currentTime - this.startTime) / this.options.duration;
+
+            if (timeFraction > 1) {
+              timeFraction = 1;
+              this.isFinished = true;
+            }
+
+            if (timeFraction <= 1) {
+              const progress = this.options.easing(timeFraction);
+
+              this.options.func(progress, {
+                startTime: this.startTime,
+                currentTime,
+                isFinished: this.isFinished,
+                options,
+              });
+
+              this.lastFrameTime = currentTime;
+            }
+
+            if (this.isFinished) {
+              this.stop();
+
+              if (typeof this.options.callback === "function") {
+                this.options.callback();
+              }
+
+              return;
+            }
+          }
+
+          this.requestId = requestAnimationFrame(animateFrame);
+        };
+      }
+
+      this.requestId = requestAnimationFrame(animateFrame);
+    }, this.options.delay);
+  }
+
+  restart() {
+    this.start();
+  }
+
+  stop() {
+    if (this.timeoutId) {
+      clearTimeout(this.timeoutId);
+    }
+
+    if (this.requestId) {
+      cancelAnimationFrame(this.requestId);
+    }
+  }
+}
