--- source/js/animation/main.js
@@ -1,5 +1,0 @@
-import Scene2DSeaCalf from "./scene-2d-seacalf";
-import Scene2dCrocodile from "./scene-2d-crocodile";
-
-export const runWinningScene = () => new Scene2DSeaCalf();
-export const runLosingScene = () => new Scene2dCrocodile();
